Move test entity to a dedicated folder
